Rename Table component to Plays
