+++ src/App.tsx
@@ -4,7 +4,7 @@ import Container from 'react-bootstrap/Container';
 import Topnav from './components/Topnav';
 import Home from './components/Home';
 import Id from './components/Id';
-import Table from './components/Table';
+import Plays from './components/Plays';
 import Details from './components/Details';
 import OriginalDetails from './components/OriginalDetails';
 import About from './components/About';
@@ -22,7 +22,7 @@ function App () {
         <Routes>
           <Route path="/" element={<Home/>} />
           <Route path="/id/:id" element={<Id/>} />
-          <Route path="/plays" element={<Table/>} />
+          <Route path="/plays" element={<Plays/>} />
           <Route path="/about" element={<About/>} />
           <Route path="/locations" element={<Map/>} />
           <Route path="/originals/:slug" element={<OriginalDetails/>} />
